docs(progress-tracker): add playground example
