+++ skills/carbon-react/components/progress-tracker.md
@@ -44,139 +44,4 @@ description: Carbon ProgressTracker component props and usage examples.
 | error | boolean \| undefined | No |  | Yes | Please use variant="error" instead. | Flag to control error state. |  |
 
 ## Examples
-### Default
-
-**Args**
-
-```tsx
-{
-    progress: 50,
-  }
-```
-
-**Render**
-
-```tsx
-(args) => <ProgressTracker {...args} />
-```
-
-
-### WithDescription
-
-**Args**
-
-```tsx
-{
-    ...Default.args,
-    description: "Description",
-  }
-```
-
-
-### CustomLabelValues
-
-**Args**
-
-```tsx
-{
-    currentProgressLabel: "£75",
-    maxProgressLabel: "£200",
-    customValuePreposition: "out of",
-    progress: Math.round((75 / 200) * 100),
-  }
-```
-
-**Render**
-
-```tsx
-(args) => <ProgressTracker {...args} />
-```
-
-
-### CustomLength
-
-**Args**
-
-```tsx
-{
-    ...Default.args,
-    length: "500px",
-  }
-```
-
-
-### LabelsPosition
-
-**Args**
-
-```tsx
-{
-    progress: 50,
-    currentProgressLabel: "50%",
-  }
-```
-
-**Render**
-
-```tsx
-(args) => (
-    <>
-      <ProgressTracker labelsPosition="top" description="Top" {...args} />
-      <ProgressTracker labelsPosition="bottom" description="Bottom" {...args} />
-      <ProgressTracker labelsPosition="left" description="Left" {...args} />
-    </>
-  )
-```
-
-
-### Sizes
-
-**Args**
-
-```tsx
-{
-    progress: 50,
-  }
-```
-
-**Render**
-
-```tsx
-(args) => (
-    <>
-      <ProgressTracker size="small" description="Small" {...args} />
-      <ProgressTracker size="medium" description="Medium" {...args} />
-      <ProgressTracker size="large" description="Large" {...args} />
-    </>
-  )
-```
-
-
-### Variants
-
-**Args**
-
-```tsx
-{
-    progress: 50,
-  }
-```
-
-**Render**
-
-```tsx
-(args) => (
-    <>
-      <ProgressTracker variant="neutral" description="Neutral" {...args} />
-      <ProgressTracker variant="warning" description="Warning" {...args} />
-      <ProgressTracker
-        variant="information"
-        description="Information"
-        {...args}
-      />
-      <ProgressTracker variant="error" description="Error" {...args} />
-      <ProgressTracker variant="success" description="Success" {...args} />
-    </>
-  )
-```
-
+No Storybook examples found.
